docs(adr006): Record containment boundary decision
Reduce the provider proof to detached-process containment and retain broker execution after native context cleanup and turn-scoped lifecycle evidence fail their gates.
