+++ fleet-rlm-implementation-plan-2026-09-06-v2.md
@@ -229,9 +229,16 @@ Do not add a permanent framework solely to collect these numbers; a small script
 
 # Phase 1 - Decide the execution containment boundary
 
-The existing live Phase 3 evidence is decisive: deleting a Daytona interpreter context did not contain a deliberately detached subprocess. A reusable Session Sandbox therefore cannot be promoted merely because the interpreter context is fresh.
+**Status: complete (2026-09-10).** The minimal provider proof repeated four
+negative results for `code_interpreter.delete_context`: an ordinary child
+completed, but a detached process-session child survived context deletion. The
+whole-Sandbox benchmark completed all deletion checks but measured 33.296 s p95
+create-through-first-execution, above the 10 s operational threshold. Retained
+broker execution (option B) remains the sole production execution boundary.
+The bounded receipts and decision rationale are recorded in the ADR006 status
+ledger.
 
-## P1.1 - Reduce the native containment proof to one minimal provider test
+## P1.1 - Reduce the native containment proof to one minimal provider test — complete
 
 **Rationale:** One small, authoritative live test is easier to reason about than containment behavior spread across a broad feasibility campaign.
 
@@ -248,7 +255,7 @@ Do not add additional fake layers around this proof.
 
 **Done when:** the provider behavior is reproducible and the result is independent of Fleet's higher-level RLM machinery.
 
-## P1.2 - Test exactly one provider-supported process-tree containment mechanism
+## P1.2 - Test exactly one provider-supported process-tree containment mechanism — complete
 
 **Rationale:** The clean native target only works if Fleet can terminate everything generated by one Run before reusing the Sandbox.
 
@@ -258,7 +265,7 @@ Do not add additional fake layers around this proof.
 
 **Done when:** either process-tree containment is demonstrated or the provider mechanism is conclusively insufficient for Fleet's contract.
 
-## P1.3 - If native context containment fails, compare the two real fallbacks
+## P1.3 - If native context containment fails, compare the two real fallbacks — complete
 
 **Rationale:** Fleet should choose one safe execution boundary rather than carrying broker and native paths indefinitely.
 
@@ -280,7 +287,7 @@ Use the existing lifecycle benchmark/evaluation harness. Do not invent a third a
 
 **Done when:** one fallback is clearly safer and operationally acceptable.
 
-## P1.4 - Record the production execution decision
+## P1.4 - Record the production execution decision — complete
 
 **Rationale:** Every later deletion depends on knowing which execution boundary is authoritative.
 
@@ -298,7 +305,7 @@ State why the other two are rejected/deferred and what rollback means.
 
 **Done when:** there is one target execution model and no unresolved "maybe both" production design.
 
-**Phase 1 exit:** containment decision made. No Phase 2 deletion begins before this decision.
+**Phase 1 exit: complete.** Retained broker execution is selected. No Phase 2 deletion begins before this decision.
 
 ---
